split file and reorganise in folders
